refactor(frontend): organize flat components into domain folders (#347)
Move Phase A domain clusters into kebab-case folders and update imports so related UI is easier to find without behavior changes.
