+++ apps/frontend/src/components/__tests__/CharacterList.test.tsx
@@ -2,7 +2,7 @@ import { describe, it, expect, vi } from "vitest";
 import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import type { Character } from "@branchforge/shared";
-import { CharacterList } from "../CharacterList";
+import { CharacterList } from "../characters/CharacterList";
 
 const mockCharacters: Character[] = [
   {
