+++ apps/frontend/src/components/characters/CharacterList.tsx
