+++ apps/frontend/src/components/ide-shared/ZipImportFilesDialog/ZipImportFilesDialog.tsx
@@ -8,7 +8,7 @@
 import { X, Package } from "lucide-react";
 import { Button } from "@/components/ui/button";
 import { Dialog, DialogContent, DialogTitle } from "@/components/ui/dialog";
-import { CharacterImportWizard } from "@/components/CharacterImportWizard.lazy";
+import { CharacterImportWizard } from "@/components/CharacterImportWizard/CharacterImportWizard.lazy";
 import { useZipImportFilesDialog } from "./useZipImportFilesDialog";
 import { ZipImportFilesDialogDropZone } from "./ZipImportFilesDialogDropZone";
 import {
