+++ apps/frontend/src/components/script-mode/GitLabSyncDialog/GitLabSyncDialog.tsx
@@ -13,7 +13,7 @@ import { useGitLabSync } from "@/hooks/useGitLabSync";
 import { useToast } from "@/contexts/ToastContext";
 import { useLabels } from "@/hooks/useLabels";
 import { characterKeys, projectFilesKeys } from "@/lib/query-keys";
-import { CharacterImportWizard } from "@/components/CharacterImportWizard.lazy";
+import { CharacterImportWizard } from "@/components/CharacterImportWizard/CharacterImportWizard.lazy";
 import { charactersApi } from "@/lib/api/characters";
 import { GitLabSyncDialogProgress } from "./GitLabSyncDialogProgress";
 import { GitLabSyncSyncForm } from "./GitLabSyncSyncForm";
