+++ apps/frontend/src/components/stats/StatEditDialog.tsx
