+++ apps/frontend/src/components/stats/StatList.tsx
