+++ apps/frontend/src/components/stats/StatProgression.tsx
