+++ apps/frontend/src/components/variables/VariableEditDialog.tsx
