+++ apps/frontend/src/components/variables/VariablesList.tsx
