+++ apps/frontend/src/components/visual-system/VisualSystemNamingTemplate.tsx
