+++ apps/frontend/src/components/visual-system/VisualSystemOptionalInputs.tsx
