+++ apps/frontend/src/components/visual-system/VisualSystemPreviewPanel.tsx
