+++ apps/frontend/src/components/visual-system/visual-system.helpers.ts
