+++ apps/frontend/src/components/world-elements/WorldElementEditDialog.tsx
